refactor: restructure TUI crate using The Elm Architecture (TEA)
